Split index.ts apart into helper files
index.ts is now almost 10% its original size and critical parts of the bot are much easier to find
Also clean up some messy imports
