+++ util/commands.ts
@@ -1,4 +1,5 @@
-import { commands, config } from "..";
+import { config } from "..";
+import { commands } from "../helpers/modules";
 import { Command } from "../Types";
 
 export const getCommand = (trigger:string) => commands.find(cmd => trigger == cmd.name || (cmd.aliases && cmd.aliases.includes(trigger)));
